moves shared out of src
